requirements.txt
python-pptx>=0.6.21    # PowerPoint file manipulation
openai>=1.0.0          # OpenAI API client
markdown>=3.4.0        # Markdown parsing
Pillow>=10.0.0        # Image processing
click>=8.1.0          # CLI framework
pydantic>=2.0.0       # Data validation
python-dotenv>=1.0.0  # Environment configuration
